requirements.txt
streamlit
pytesseract
pillow
